+++ app/process.py
@@ -77,6 +77,7 @@ def analyse(
             imagery, imagery_size = imagery_store.imagery(area_coords, start_date, end_date)
             imagery = tx({'imagery': imagery})
 
+            log.debug('Starting inference')
             logits = inference_session.run(output_names=None, input_feed=imagery)[0][0]
             labels = __fusion(osm, osm_lulc_mapping, threshold, area_coords, end_date, fusion_mode, logits)